Remplacement long integer

+++ loan-broker/src/main/java/fr/pantheonsorbonne/urf27/miage/model/Borrower.java
@@ -10,7 +10,7 @@ public class Borrower {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idBorrower", nullable = false)
-    private long idBorrower;
+    private Integer idBorrower;
 
     @Column(name = "name", nullable = false)
     private String name;
